refactor: CQRS 파일 구조 사용하지 않도록 수정 - #303
- 아직 해당 파일들이 CQRS로 구조가 분리되지 않았기 때문에 파일 구조 원복
