+++ product/src/main/java/com/back/product/app/facade/ProductFacade.java
@@ -5,19 +5,15 @@ import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
 import com.back.product.adapter.out.event.BrandSpringEventPublisher;
 import com.back.product.adapter.out.event.ProductSpringEventPublisher;
-import com.back.product.app.usecase.command.*;
+import com.back.product.app.usecase.*;
 import com.back.product.domain.*;
 import com.back.product.dto.command.*;
 import com.back.product.dto.model.*;
 import com.back.product.dto.request.*;
 import com.back.product.dto.response.*;
-import com.back.product.event.spring.ProductCreationCompletedEvent;
-import com.back.product.event.spring.ProductDeletionCompletedEvent;
-import com.back.product.event.spring.ProductUpdateCompletedEvent;
 import com.back.product.mapper.*;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
