+++ product/src/main/java/com/back/product/app/usecase/OptionUseCase.java
@@ -1,27 +1,19 @@
-package com.back.product.app.usecase.command;
+package com.back.product.app.usecase;
 
 import com.back.common.annotation.Loggable;
 import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
 import com.back.product.adapter.out.persistence.OptionGroupRepository;
 import com.back.product.adapter.out.persistence.OptionValueRepository;
-import com.back.product.app.usecase.query.ProductSupport;
 import com.back.product.domain.OptionGroup;
 import com.back.product.domain.OptionValue;
 import com.back.product.dto.command.OptionCreateCommand;
 import com.back.product.dto.model.OptionDto;
-import com.back.product.dto.request.OptionAppendRequestDto;
-import com.back.product.dto.request.OptionListCreateRequestDto;
-import com.back.product.dto.request.OptionGroupModifyRequestDto;
-import com.back.product.dto.request.OptionValueModifyRequestDto;
 import com.back.product.dto.response.OptionGroupModifyResponseDto;
-import com.back.product.dto.response.OptionListResponseDto;
-import com.back.product.dto.response.OptionResponseDto;
 import com.back.product.dto.response.OptionValueModifyResponseDto;
 import com.back.product.mapper.OptionGroupMapper;
 import com.back.product.mapper.OptionMapper;
 import com.back.product.mapper.OptionValueMapper;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
