+++ product/src/test/java/com/back/product/app/usecase/CategoryUseCaseTest.java
@@ -3,13 +3,9 @@ package com.back.product.app.usecase;
 import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
 import com.back.product.adapter.out.persistence.CategoryRepository;
-import com.back.product.app.usecase.command.CategoryUseCase;
-import com.back.product.app.usecase.query.ProductSupport;
 import com.back.product.domain.Category;
 import com.back.product.dto.command.CategoryDataCommand;
 import com.back.product.dto.model.CategoryDto;
-import com.back.product.dto.request.CategoryListCreateRequestDto;
-import com.back.product.dto.request.CategoryDataRequestDto;
 import com.back.product.mapper.CategoryMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
