+++ product/src/test/java/com/back/product/app/usecase/OptionUseCaseTest.java
@@ -4,8 +4,6 @@ import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
 import com.back.product.adapter.out.persistence.OptionGroupRepository;
 import com.back.product.adapter.out.persistence.OptionValueRepository;
-import com.back.product.app.usecase.command.OptionUseCase;
-import com.back.product.app.usecase.query.ProductSupport;
 import com.back.product.domain.OptionGroup;
 import com.back.product.domain.OptionValue;
 import com.back.product.dto.command.OptionCreateCommand;
