+++ product/src/test/java/com/back/product/app/usecase/ProductInfoUseCaseTest.java
@@ -3,13 +3,10 @@ package com.back.product.app.usecase;
 import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
 import com.back.product.adapter.out.persistence.ProductInfoRepository;
-import com.back.product.app.usecase.command.ProductInfoUseCase;
-import com.back.product.app.usecase.query.ProductSupport;
 import com.back.product.domain.Brand;
 import com.back.product.domain.Category;
 import com.back.product.domain.ProductInfo;
 import com.back.product.dto.command.ProductInfoDataCommand;
-import com.back.product.dto.request.ProductInfoDataRequestDto;
 import com.back.product.mapper.ProductInfoMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
